+++ src/graphics/VertexArray.h
@@ -12,6 +12,9 @@
 
 namespace Graphics {
 
+	struct VertexFormatDesc;
+	class VertexBuffer;
+
 	/*
  * VertexArray is a multi-purpose vertex container. Users specify
  * the attributes they intend to use and then add vertices. Renderers
@@ -35,6 +38,20 @@ namespace Graphics {
 		//removes vertices, does not deallocate space
 		// if reserveSize != 0, ensures the array has enough space for at least that many elements
 		void Clear(uint32_t reserveSize = 0);
+
+		// Copy the vertices from this VertexArray into the given VertexBuffer.
+		// The VertexBuffer must have a compatible vertex format (i.e. stride).
+		// NOTE: only supports these formats:
+		//  - POSITION | [DIFFUSE/NORMAL/UV0]
+		//  - POSITION | DIFFUSE | [UV0]
+		//  - POSITION | NORMAL  | [UV0/DIFFUSE]
+		bool Populate(VertexBuffer *buffer) const;
+
+		// Copy the vertices from this VertexArray into the given buffer.
+		// Supports any combination of attributes, but is slightly slower than Populate().
+		// Returns false if the passed buffer cannot hold this VertexArray.
+		// Note that the passed fmt must have the same order of attributes as one created from VertexFormatDesc::FromAttribSet
+		bool PopulateRange(const VertexFormatDesc &fmt, uint8_t *buffer, size_t size) const;
 
 		// don't mix these
 		void Add(const vector3f &v);
