+++ src/graphics/opengl/VertexBufferGL.cpp
@@ -118,6 +118,11 @@ namespace Graphics {
 			delete[] m_data;
 		}
 
+		bool VertexBuffer::Populate(const VertexArray &va, const VertexFormatDesc &)
+		{
+			va.Populate(this);
+		}
+
 		uint8_t *VertexBuffer::MapInternal(BufferMapMode mode)
 		{
 			PROFILE_SCOPED()
@@ -208,166 +213,6 @@ namespace Graphics {
 			glBindBuffer(GL_ARRAY_BUFFER, 0);
 		}
 
-#pragma pack(push, 4)
-		struct PosUVVert {
-			vector3f pos;
-			vector2f uv;
-		};
-
-		struct PosNormVert {
-			vector3f pos;
-			vector3f norm;
-		};
-
-		struct PosColVert {
-			vector3f pos;
-			Color4ub col;
-		};
-
-		struct PosVert {
-			vector3f pos;
-		};
-
-		struct PosColUVVert {
-			vector3f pos;
-			Color4ub col;
-			vector2f uv;
-		};
-
-		struct PosNormUVVert {
-			vector3f pos;
-			vector3f norm;
-			vector2f uv;
-		};
-
-		struct PosNormColVert {
-			vector3f pos;
-			vector3f norm;
-			Color4ub col;
-		};
-#pragma pack(pop)
-
-		static inline void CopyPosNorm(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosNormVert *vtxPtr = vb->Map<PosNormVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosNormVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].norm = va.normal[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPosUV0(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosUVVert *vtxPtr = vb->Map<PosUVVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosUVVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].uv = va.uv0[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPosCol(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosColVert *vtxPtr = vb->Map<PosColVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosColVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].col = va.diffuse[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPos(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosVert *vtxPtr = vb->Map<PosVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPosColUV0(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosColUVVert *vtxPtr = vb->Map<PosColUVVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosColUVVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].col = va.diffuse[i];
-				vtxPtr[i].uv = va.uv0[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPosNormUV0(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosNormUVVert *vtxPtr = vb->Map<PosNormUVVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosNormUVVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].norm = va.normal[i];
-				vtxPtr[i].uv = va.uv0[i];
-			}
-			vb->Unmap();
-		}
-
-		static inline void CopyPosNormCol(Graphics::VertexBuffer *vb, const Graphics::VertexArray &va)
-		{
-			PosNormColVert *vtxPtr = vb->Map<PosNormColVert>(Graphics::BUFFER_MAP_WRITE);
-			assert(vb->GetDesc().stride == sizeof(PosNormColVert));
-			for (Uint32 i = 0; i < va.GetNumVerts(); i++) {
-				vtxPtr[i].pos = va.position[i];
-				vtxPtr[i].norm = va.normal[i];
-				vtxPtr[i].col = va.diffuse[i];
-			}
-			vb->Unmap();
-		}
-
-		// copies the contents of the VertexArray into the buffer
-		bool VertexBuffer::Populate(const VertexArray &va, const VertexFormatDesc &)
-		{
-			PROFILE_SCOPED()
-			assert(va.GetNumVerts() > 0);
-			assert(va.GetNumVerts() <= m_capacity);
-			bool result = false;
-			const Graphics::AttributeSet as = va.GetAttributeSet();
-			switch (as) {
-			case Graphics::ATTRIB_POSITION:
-				CopyPos(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_DIFFUSE:
-				CopyPosCol(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_NORMAL:
-				CopyPosNorm(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_UV0:
-				CopyPosUV0(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_DIFFUSE | Graphics::ATTRIB_UV0:
-				CopyPosColUV0(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_NORMAL | Graphics::ATTRIB_UV0:
-				CopyPosNormUV0(this, va);
-				result = true;
-				break;
-			case Graphics::ATTRIB_POSITION | Graphics::ATTRIB_NORMAL | Graphics::ATTRIB_DIFFUSE:
-				CopyPosNormCol(this, va);
-				result = true;
-				break;
-			}
-			SetVertexCount(va.GetNumVerts());
-			return result;
-		}
-
 		void VertexBuffer::BufferData(const size_t size, void *data)
 		{
 			PROFILE_SCOPED()
@@ -398,47 +243,12 @@ namespace Graphics {
 			m_lastFlushed = 0;
 		}
 
-		// TODO: consider moving Populate to VertexArray and providing a MapRange/FlushRange API for regular VertexBuffer?
 		bool CachedVertexBuffer::Populate(const VertexArray &va, const VertexFormatDesc &fmt)
 		{
-			assert(m_capacity - m_size >= va.GetNumVerts());
-
-			// Get an enumerated list of the attributes in this set
-			VertexAttrib semantics[MAX_ATTRIBS] = {};
-
-			uint32_t numAttrs = fmt.GetNumAttribs();
-			GetActiveAttribsInSet(va.GetAttributeSet(), semantics, numAttrs);
-
-			// Complicated-ish loop to deal with 16+ possible combinations of vertex formats
-			// (Position is effectively required, or it would be 32)
-			for (size_t idx = 0; idx < va.GetNumVerts(); idx++) {
-				for (uint32_t n = 0; n < numAttrs; n++) {
-					// Calculate the location of this component inside the vertex being written
-					uint8_t *data = m_data + (m_size + idx) * m_desc.stride + fmt.attribs[n].offset;
-					switch (semantics[n]) {
-					case ATTRIB_POSITION:
-						*reinterpret_cast<vector3f *>(data) = va.position[idx];
-						break;
-					case ATTRIB_NORMAL:
-						*reinterpret_cast<vector3f *>(data) = va.normal[idx];
-						break;
-					case ATTRIB_DIFFUSE:
-						*reinterpret_cast<Color4ub *>(data) = va.diffuse[idx];
-						break;
-					case ATTRIB_UV0:
-						*reinterpret_cast<vector2f *>(data) = va.uv0[idx];
-						break;
-					case ATTRIB_TANGENT:
-						*reinterpret_cast<vector3f *>(data) = va.tangent[idx];
-						break;
-					default:
-						assert(false && "Unimplemented vertex attribute in CachedVertexBuffer::Populate!");
-						break;
-					}
-				}
-			}
-
+			assert(fmt.bindings[0].stride == m_desc.stride);
+			va.PopulateRange(fmt, m_data + m_size * m_desc.stride, (m_capacity - m_size) * m_desc.stride);
 			m_size += va.GetNumVerts();
+
 			return true;
 		}
